Add files via upload
This utility was written for use with the Ronoth LoStik LoRa transceiver.  It is intended to be run on Linux but can be adapted for Windows with some modification.  The utility connects to the LoStik via its serial interface and polls all of the current LoRa mode configuration parameters.  Also presented are the device defaults as described within the Microchip RN2903 command reference document.