Add files via upload
This utility was written for use with the Ronoth LoStik LoRa transceiver.  It is intended to be run on Linux but can be adapted for Windows with some modification.  The utility connects to the LoStik via its serial interface and writes all desired LoRa configuration parameters to the device.  The device will generally respond with 'ok' when a parameter is successfully written.